ISAPP-67: codestyle + swagger
