+++ src/main/java/vsu/cs/is/infsysserver/employee/adapter/rest/EmployeeController.java
@@ -11,8 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import vsu.cs.is.infsysserver.employee.EmployeeServiceImplementation;
-import vsu.cs.is.infsysserver.employee.EmployeeServiceMock;
+import vsu.cs.is.infsysserver.employee.EmployeeService;
 import vsu.cs.is.infsysserver.employee.adapter.rest.api.EmployeeApi;
 import vsu.cs.is.infsysserver.employee.adapter.rest.dto.request.EmployeeCreateRequest;
 import vsu.cs.is.infsysserver.employee.adapter.rest.dto.request.EmployeeUpdateRequest;
@@ -28,13 +27,12 @@ import static org.springframework.http.ResponseEntity.ok;
 @RequiredArgsConstructor
 @CrossOrigin
 public class EmployeeController implements EmployeeApi {
-    private final EmployeeServiceImplementation employeeService;
-    private final EmployeeServiceMock employeeServiceMock;
+    private final EmployeeService employeeService;
 
     @Override
     @GetMapping
     public ResponseEntity<List<EmployeeResponse>> getAllEmployees() {
-        return ok(employeeServiceMock.getAllEmployees());
+        return ok(employeeService.getAllEmployees());
     }
 
     @Override
