+++ src/main/java/vsu/cs/is/infsysserver/user/adapter/rest/UserController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import vsu.cs.is.infsysserver.employee.EmployeeServiceImplementation;
+import vsu.cs.is.infsysserver.employee.EmployeeService;
 import vsu.cs.is.infsysserver.employee.adapter.rest.dto.response.EmployeeResponse;
 import vsu.cs.is.infsysserver.user.adapter.rest.api.UserApi;
 
@@ -19,7 +19,7 @@ import static org.springframework.http.ResponseEntity.ok;
 @CrossOrigin
 public class UserController implements UserApi {
 
-    private final EmployeeServiceImplementation employeeService;
+    private final EmployeeService employeeService;
 
     @Override
     @GetMapping
